+++ packages/flutter_background_service_android/android/src/main/java/id/flutter/flutter_background_service/FlutterBackgroundServicePlugin.java
@@ -3,51 +3,90 @@ package id.flutter.flutter_background_service;
 import static android.content.Context.MODE_PRIVATE;
 
 import android.app.ActivityManager;
-import android.content.BroadcastReceiver;
+import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
+import android.content.ServiceConnection;
 import android.content.SharedPreferences;
+import android.os.Handler;
+import android.os.IBinder;
+import android.os.RemoteException;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
-import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
-import org.json.JSONException;
 import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.service.ServiceAware;
 import io.flutter.embedding.engine.plugins.service.ServicePluginBinding;
+import io.flutter.plugin.common.JSONMethodCodec;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
-import io.flutter.plugin.common.JSONMethodCodec;
 
 /** FlutterBackgroundServicePlugin */
-public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements FlutterPlugin, MethodCallHandler, ServiceAware {
+public class FlutterBackgroundServicePlugin implements FlutterPlugin, MethodCallHandler, ServiceAware {
   private static final String TAG = "BackgroundServicePlugin";
-  private static final List<FlutterBackgroundServicePlugin> _instances = new ArrayList<>();
-
+  private Handler mainHandler;
   public FlutterBackgroundServicePlugin() {
-    _instances.add(this);
+
   }
 
   private MethodChannel channel;
   private Context context;
-  private BackgroundService service;
+
+  private IBackgroundServiceBinder serviceBinder;
+  private final int binderId = (int) (System.currentTimeMillis()/1000);
+  private final ServiceConnection serviceConnection = new ServiceConnection() {
+    @Override
+    public void onServiceConnected(ComponentName name, IBinder service) {
+      serviceBinder = IBackgroundServiceBinder.Stub.asInterface(service);
+
+      try {
+        IBackgroundService listener = new IBackgroundService.Stub() {
+          @Override
+          public void invoke(String data) throws RemoteException {
+            try {
+              JSONObject call = new JSONObject(data);
+              receiveData(call);
+            }catch (Exception e){
+              e.printStackTrace();
+            }
+          }
+
+          @Override
+          public void stop() throws  RemoteException {
+            if (context != null && serviceBinder != null){
+              context.unbindService(serviceConnection);
+            }
+          }
+        };
+
+        serviceBinder.bind(binderId, listener);
+      }catch (Exception e){
+        e.printStackTrace();
+      }
+    }
+
+    @Override
+    public void onServiceDisconnected(ComponentName name) {
+      try {
+        serviceBinder.unbind(binderId);
+        serviceBinder = null;
+      }catch (Exception e){
+        e.printStackTrace();
+      }
+    }
+  };
 
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
     this.context = flutterPluginBinding.getApplicationContext();
-    LocalBroadcastManager localBroadcastManager = LocalBroadcastManager.getInstance(this.context);
-    localBroadcastManager.registerReceiver(this, new IntentFilter("id.flutter/background_service"));
+    mainHandler = new Handler(context.getMainLooper());
 
     channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "id.flutter/background_service_android", JSONMethodCodec.INSTANCE);
     channel.setMethodCallHandler(this);
@@ -55,22 +94,22 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
 
   @SuppressWarnings("deprecation")
   public static void registerWith(Registrar registrar) {
-    LocalBroadcastManager localBroadcastManager = LocalBroadcastManager.getInstance(registrar.context());
     final FlutterBackgroundServicePlugin plugin = new FlutterBackgroundServicePlugin();
-    localBroadcastManager.registerReceiver(plugin, new IntentFilter("id.flutter/background_service"));
-
     final MethodChannel channel = new MethodChannel(registrar.messenger(), "id.flutter/background_service_android", JSONMethodCodec.INSTANCE);
     channel.setMethodCallHandler(plugin);
     plugin.channel = channel;
   }
 
-  private static void configure(Context context, long entrypointHandle, long backgroundHandle, boolean isForeground, boolean autoStartOnBoot) {
+  private static void configure(Context context, long backgroundHandle, boolean isForeground, boolean autoStartOnBoot, String initialNotificationContent, String initialNotificationTitle, String notificationChannelId, int foregroundNotificationId) {
     SharedPreferences pref = context.getSharedPreferences("id.flutter.background_service", MODE_PRIVATE);
     pref.edit()
-            .putLong("entrypoint_handle", entrypointHandle)
             .putLong("background_handle", backgroundHandle)
             .putBoolean("is_foreground", isForeground)
             .putBoolean("auto_start_on_boot", autoStartOnBoot)
+            .putString("initial_notification_content", initialNotificationContent)
+            .putString("initial_notification_title", initialNotificationTitle)
+            .putString("notification_channel_id", notificationChannelId)
+            .putInt("foreground_notification_id", foregroundNotificationId)
             .apply();
   }
 
@@ -78,11 +117,15 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
     BackgroundService.enqueue(context);
     boolean isForeground = BackgroundService.isForegroundService(context);
     Intent intent = new Intent(context, BackgroundService.class);
+    intent.putExtra("binder_id", binderId);
+
     if (isForeground) {
       ContextCompat.startForegroundService(context, intent);
     } else {
       context.startService(intent);
     }
+
+    context.bindService(intent, serviceConnection, Context.BIND_AUTO_CREATE);
   }
 
   @Override
@@ -92,12 +135,15 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
 
     try {
       if ("configure".equals(method)) {
-        long entrypointHandle = arg.getLong("entrypoint_handle");
         long backgroundHandle = arg.getLong("background_handle");
         boolean isForeground = arg.getBoolean("is_foreground_mode");
         boolean autoStartOnBoot = arg.getBoolean("auto_start_on_boot");
+        String initialNotificationTitle = arg.getString("initial_notification_title");
+        String initialNotificationContent = arg.getString("initial_notification_content");
+        String notificationChannelId = arg.isNull("notification_channel_id") ? null : arg.getString("notification_channel_id");
+        int foregroundNotificationId = arg.getInt("foreground_notification_id");
 
-        configure(context, entrypointHandle, backgroundHandle, isForeground, autoStartOnBoot);
+        configure(context, backgroundHandle, isForeground, autoStartOnBoot, initialNotificationContent, initialNotificationTitle, notificationChannelId, foregroundNotificationId);
         if (autoStartOnBoot) {
           start();
         }
@@ -113,14 +159,13 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
       }
 
       if (method.equalsIgnoreCase("sendData")) {
-        for (FlutterBackgroundServicePlugin plugin : _instances) {
-          if (plugin.service != null) {
-            plugin.service.receiveData((JSONObject) call.arguments);
-            break;
-          }
+        if (serviceBinder != null) {
+          serviceBinder.invoke(call.arguments.toString());
+          result.success(true);
+          return;
         }
 
-        result.success(true);
+        result.success(false);
         return;
       }
 
@@ -135,7 +180,6 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
     }
   }
 
-  @SuppressWarnings("deprecation")
   private boolean isServiceRunning() {
     ActivityManager manager = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
     for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
@@ -149,40 +193,31 @@ public class FlutterBackgroundServicePlugin extends BroadcastReceiver implements
   @Override
   public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     channel.setMethodCallHandler(null);
+    channel = null;
 
-    LocalBroadcastManager localBroadcastManager = LocalBroadcastManager.getInstance(this.context);
-    localBroadcastManager.unregisterReceiver(this);
+    if (serviceBinder != null){
+      binding.getApplicationContext().unbindService(serviceConnection);
+    }
   }
 
-  @Override
-  public void onReceive(Context context, Intent intent) {
-    if (intent.getAction() == null) return;
-
-    if (intent.getAction().equalsIgnoreCase("id.flutter/background_service")) {
-      String data = intent.getStringExtra("data");
-      try {
-        JSONObject jData = new JSONObject(data);
-        if (channel != null) {
-          channel.invokeMethod("onReceiveData", jData);
+  private void receiveData(JSONObject data){
+    mainHandler.post(new Runnable() {
+      @Override
+      public void run() {
+        if (channel != null){
+          channel.invokeMethod("onReceiveData", data);
         }
-      } catch (JSONException e) {
-        e.printStackTrace();
-      } catch (Exception e) {
-        e.printStackTrace();
       }
-    }
+    });
   }
 
   @Override
   public void onAttachedToService(@NonNull ServicePluginBinding binding) {
     Log.d(TAG, "onAttachedToService");
-
-    this.service = (BackgroundService) binding.getService();
   }
 
   @Override
   public void onDetachedFromService() {
-    this.service = null;
     Log.d(TAG, "onDetachedFromService");
   }
 }