+++ packages/flutter_background_service_ios/lib/flutter_background_service_ios.dart
@@ -1,3 +1,5 @@
+library flutter_background_service_ios;
+
 import 'dart:async';
 import 'dart:ui';
 
@@ -5,7 +7,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_background_service_platform_interface/flutter_background_service_platform_interface.dart';
 
-Future<void> _foregroundEntrypoint() async {
+@pragma('vm:entry-point')
+Future<void> foregroundEntrypoint() async {
   WidgetsFlutterBinding.ensureInitialized();
   final service = IOSServiceInstance._();
   final int handle = await service._getForegroundHandler();
@@ -16,7 +19,8 @@ Future<void> _foregroundEntrypoint() async {
   }
 }
 
-Future<void> _backgroundEntrypoint() async {
+@pragma('vm:entry-point')
+Future<void> backgroundEntrypoint() async {
   WidgetsFlutterBinding.ensureInitialized();
   final service = IOSServiceInstance._();
   final int handle = await service._getBackgroundHandler();
@@ -63,18 +67,6 @@ class FlutterBackgroundServiceIOS extends FlutterBackgroundServicePlatform {
   }) async {
     _channel.setMethodCallHandler(_handle);
 
-    final CallbackHandle? foregroundEntrypointHandle =
-        PluginUtilities.getCallbackHandle(_foregroundEntrypoint);
-    if (foregroundEntrypointHandle == null) {
-      return false;
-    }
-
-    final CallbackHandle? backgroundEntrypointHandle =
-        PluginUtilities.getCallbackHandle(_backgroundEntrypoint);
-    if (backgroundEntrypointHandle == null) {
-      return false;
-    }
-
     final CallbackHandle? foregroundHandle =
         PluginUtilities.getCallbackHandle(iosConfiguration.onForeground);
     if (foregroundHandle == null) {
@@ -91,10 +83,6 @@ class FlutterBackgroundServiceIOS extends FlutterBackgroundServicePlatform {
     final result = await _channel.invokeMethod(
       "configure",
       {
-        "foreground_entrypoint_handle":
-            foregroundEntrypointHandle.toRawHandle(),
-        "background_entrypoint_handle":
-            backgroundEntrypointHandle.toRawHandle(),
         "background_handle": backgroundHandle.toRawHandle(),
         "foreground_handle": foregroundHandle.toRawHandle(),
         "auto_start": iosConfiguration.autoStart,