+++ packages/flutter_background_service_platform_interface/lib/src/configs.dart
@@ -33,14 +33,25 @@ class AndroidConfiguration {
   /// whether service is foreground or background mode
   final bool isForegroundMode;
 
-  final String? foregroundServiceNotificationTitle;
-  final String? foregroundServiceNotificationContent;
+  /// notification content that will be shown on status bar when the background service is starting
+  /// defaults to "Preparing"
+  final String initialNotificationContent;
+  final String initialNotificationTitle;
+
+  /// use custom notification channel id
+  /// you must to create the notification channel before you run configure() method.
+  final String? notificationChannelId;
+
+  /// notification id will be used by foreground service
+  final int foregroundServiceNotificationId;
 
   AndroidConfiguration({
     required this.onStart,
     this.autoStart = true,
     required this.isForegroundMode,
-    this.foregroundServiceNotificationContent,
-    this.foregroundServiceNotificationTitle,
+    this.initialNotificationContent = 'Preparing',
+    this.initialNotificationTitle = 'Background Service',
+    this.notificationChannelId,
+    this.foregroundServiceNotificationId = 112233,
   });
 }